third_party/WebKit/LayoutTests/inspector/elements/elements-panel-correct-case.html

<!DOCTYPE html>
<html>
<head>
<link rel="stylesheet" type="text/css" href="resources/elements-panel-styles.css">

<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>
function test()
{
    ElementsTestRunner.expandElementsTree(step1);

    function step1()
    {
        ElementsTestRunner.dumpElementsTree();
        TestRunner.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p>
Tests that elements panel shows all types of elements in the correct case.
</p>
<svg>
    <feComposite></feComposite>
</svg>
<svg>
    <circle></circle>
</svg>
<DIv></DIv>
<DIV></DIV>
</body>
</html>
